packages/workers/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
mysql-connector-python>=8.1.0

# Async I/O
aiohttp>=3.9.0

# Redis
redis>=4.5.0
redis[async]>=4.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Logging
python-json-logger>=2.0.0

# Data validation
pydantic>=2.0.0

# Retry logic
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Jina AI embeddings
jina>=0.8.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.4.0
flake8>=6.0.0
